legger til tester for å sjekke retries
